+++ flask/helpers.py
@@ -789,19 +789,38 @@ def _matching_loader_thinks_module_is_package(loader, mod_name):
         loader.__class__.__name__)
 
 
-def find_package(import_name):
-    """Finds a package and returns the prefix (or None if the package is
-    not installed) as well as the folder that contains the package or
-    module as a tuple.  The package path returned is the module that would
-    have to be added to the pythonpath in order to make it possible to
-    import the module.  The prefix is the path below which a UNIX like
-    folder structure exists (lib, share etc.).
-    """
-    root_mod_name = import_name.split('.')[0]
+def _find_package_path(root_mod_name):
+    """Find the path where the module's root exists in"""
+    if sys.version_info >= (3, 4):
+        import importlib.util
+
+        try:
+            spec = importlib.util.find_spec(root_mod_name)
+            if spec is None:
+                raise ValueError("not found")
+        # ImportError: the machinery told us it does not exist
+        # ValueError:
+        #    - the module name was invalid
+        #    - the module name is __main__
+        #    - *we* raised `ValueError` due to `spec` being `None`
+        except (ImportError, ValueError):
+            pass  # handled below
+        else:
+            # namespace package
+            if spec.origin in {"namespace", None}:
+                return os.path.dirname(next(iter(spec.submodule_search_locations)))
+            # a package (with __init__.py)
+            elif spec.submodule_search_locations:
+                return os.path.dirname(os.path.dirname(spec.origin))
+            # just a normal module
+            else:
+                return os.path.dirname(spec.origin)
+
+    # we were unable to find the `package_path` using PEP 451 loaders
     loader = pkgutil.get_loader(root_mod_name)
-    if loader is None or import_name == '__main__':
+    if loader is None or root_mod_name == '__main__':
         # import name is not found, or interactive/main module
-        package_path = os.getcwd()
+        return os.getcwd()
     else:
         # For .egg, zipimporter does not have get_filename until Python 2.7.
         if hasattr(loader, 'get_filename'):
@@ -815,17 +834,29 @@ def find_package(import_name):
             # Google App Engine's HardenedModulesHook
             #
             # Fall back to imports.
-            __import__(import_name)
-            filename = sys.modules[import_name].__file__
+            __import__(root_mod_name)
+            filename = sys.modules[root_mod_name].__file__
         package_path = os.path.abspath(os.path.dirname(filename))
 
         # In case the root module is a package we need to chop of the
         # rightmost part.  This needs to go through a helper function
         # because of python 3.3 namespace packages.
-        if _matching_loader_thinks_module_is_package(
-                loader, root_mod_name):
+        if _matching_loader_thinks_module_is_package(loader, root_mod_name):
             package_path = os.path.dirname(package_path)
 
+    return package_path
+
+
+def find_package(import_name):
+    """Finds a package and returns the prefix (or None if the package is
+    not installed) as well as the folder that contains the package or
+    module as a tuple.  The package path returned is the module that would
+    have to be added to the pythonpath in order to make it possible to
+    import the module.  The prefix is the path below which a UNIX like
+    folder structure exists (lib, share etc.).
+    """
+    root_mod_name, _, _ = import_name.partition('.')
+    package_path = _find_package_path(root_mod_name)
     site_parent, site_folder = os.path.split(package_path)
     py_prefix = os.path.abspath(sys.prefix)
     if package_path.startswith(py_prefix):